refactor(frontend): slim the waitlist container and stop logging the payload
Two Aikido findings on the waitlist PR:
- fetchWaitlist logged the entire response (which can carry patientId and
  user-entered notes) on the not-an-array branch; log the type instead.
- WaitlistPanel combined permission checks, loading, companion mapping, action
  refreshes, error handling and rendering in one function. Move all of that into
  a useWaitlist hook so the container is a thin projection onto the panel.

Behaviour identical; the service and container tests pass unchanged (27/27),
the hook is exercised through the container test. tsc 0, lint clean.
